Move step details components into a subfolder
